requirements.txt
openai
flask
python-dotenv
pymongo
bson